+++ packages/vue-mapbox-gl/components/MapboxCluster.vue
@@ -45,9 +45,27 @@
       default: 50,
     },
     /**
+     * Minimum number of points necessary to form a cluster.
+     * @type {number}
+     */
+    clusterMinPoints: {
+      type: Number,
+      default: 2,
+    },
+    /**
+     * An object defining custom properties on the generated clusters.
+     * @see  https://docs.mapbox.com/style-spec/reference/sources/#geojson-clusterProperties
+     * @see  https://docs.mapbox.com/mapbox-gl-js/example/cluster-html/
+     * @type {object}
+     */
+    clusterProperties: {
+      type: Object,
+      default: () => ({}),
+    },
+    /**
      * The layout configuration for the clusters circles
      * @see  https://docs.mapbox.com/mapbox-gl-js/example/cluster/
-     * @type {Object}
+     * @type {object}
      */
     clustersLayout: {
       type: Object,
@@ -56,7 +74,7 @@
     /**
      * The paint configuration for the clusters circles
      * @see  https://docs.mapbox.com/mapbox-gl-js/example/cluster/
-     * @type {Object}
+     * @type {object}
      */
     clustersPaint: {
       type: Object,
@@ -68,7 +86,7 @@
     /**
      * The layout configuration for the clusters count
      * @see  https://docs.mapbox.com/mapbox-gl-js/example/cluster/
-     * @type {Object}
+     * @type {object}
      */
     clusterCountLayout: {
       type: Object,
@@ -79,7 +97,7 @@
     /**
      * The paint configuration for the clusters count
      * @see  https://docs.mapbox.com/mapbox-gl-js/example/cluster/
-     * @type {Object}
+     * @type {object}
      */
     clusterCountPaint: {
       type: Object,
@@ -99,7 +117,7 @@
     /**
      * The layout configuration for the unclustered points
      * @see  https://docs.mapbox.com/mapbox-gl-js/example/cluster/
-     * @type {Object}
+     * @type {object}
      */
     unclusteredPointLayout: {
       type: Object,
@@ -108,7 +126,7 @@
     /**
      * The paint configuration for the unclustered points
      * @see  https://docs.mapbox.com/mapbox-gl-js/example/cluster/
-     * @type {Object}
+     * @type {object}
      */
     unclusteredPointPaint: {
       type: Object,
@@ -140,13 +158,15 @@
 
   const sourceId = computed(() => getId('source'));
   const source = computed(() => {
-    const { data, clusterMaxZoom, clusterRadius } = props;
+    const { data, clusterMaxZoom, clusterRadius, clusterMinPoints, clusterProperties } = props;
     return {
       type: 'geojson',
       cluster: true,
       data,
       clusterMaxZoom,
       clusterRadius,
+      clusterMinPoints,
+      clusterProperties,
     };
   });
 
@@ -170,7 +190,7 @@
 
   /**
    * The unclustered points layer
-   * @type {Object}
+   * @type {object}
    */
   const unclusteredPointLayer = computed(() => ({
     id: getId('unclustered-point'),
@@ -183,8 +203,7 @@
 
   /**
    * Click handler for the clusters layer to zoom on the clicked cluster
-   *
-   * @param  {Object} event The Mapbox click event's object
+   * @param  {object} event The Mapbox click event's object
    * @returns {void}
    */
   function clustersClickHandler(event) {
@@ -216,7 +235,6 @@
   }
   /**
    * Mouseenter handler for the clusters layer to set a pointer cursor
-   *
    * @returns {void}
    */
   function clustersMouseenterHandler() {
@@ -224,7 +242,6 @@
   }
   /**
    * Mouseleave handler for the clusters layer to unset the pointer cursor
-   *
    * @returns {void}
    */
   function clustersMouseleaveHandler() {
@@ -234,8 +251,7 @@
   /**
    * Handler for the click event on a single feature, emits an event with
    * the feature object and the original event object
-   *
-   * @param  {Object} event The Mapbox click event's object
+   * @param  {object} event The Mapbox click event's object
    * @returns {void}
    */
   function unclusteredPointClickHandler(event) {
@@ -247,8 +263,7 @@
    * Handler for the mouseenter event on a single feature.
    * Emits an event with the feature object and the original event as
    * parameters, and sets the cursor style to pointer.
-   *
-   * @param  {Object} event The Mapbox mouseenter event's object
+   * @param  {object} event The Mapbox mouseenter event's object
    * @returns {void}
    */
   function unclusteredPointMouseenterHandler(event) {
@@ -261,8 +276,7 @@
    * Handler for the mouseleave event on a single feature.
    * Emits an event with the original event object as parameter, and resets
    * the cursor style to its default value.
-   *
-   * @param  {Object} event The Mapbox mouselvea event‘s object
+   * @param  {object} event The Mapbox mouselvea event‘s object
    * @returns {void}
    */
   function unclusteredPointMouseleaveHandler(event) {